backend/requirement.txt
web3==6.15.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
flask==3.0.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
